Lab_work_6/user_control_from_kafka_to_db/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
passlib==1.7.4